Add diamond type arguments where raw types were used when possible.
Raw `List`, `ArrayList`, `Class` and other types were used in many files. I used Eclipse and NetBeans to identify and transform these into fully typed variables/parameters/arguments and diamond-typed definitions and references.

In a few places this was not possible; I added a FIXME in one of them (in DataversePage.java).