authenticateduserlookup: use ordinary long as primary key #796
Previously, the combination of two strings was used as a primary key but
we need to be able to update those strings when converting from one type
of user to another, such as from BuiltIn to Shibboleth. Using an
ordinary primary key (a long) allows us to update the strings that used
to make up the primary key.