Indexing: better permission updates #1901
Datasets that are direct children of the root dataverse were not being
indexed when roles were granted. This should be fixed.

Because there is no inheritance of permissions, we switched away from
using a Solr query to determine the subtree of candidates for
re-indexing. We are querying the database instead.