Shib: use port 8181 for Shib page and plumbing #2117
Apache and mod_shib once ran on port 443 but we decided to let Glassfish
serve HTTPS requests there instead. Apache is still serving HTTP on port
80 but is now serving HTTPS on port 8181 to the code and configuration
has been changed to accomodate this.

Configuration file examples are from http://shibtest.dataverse.org with
comments. The idea is that the only time users see port 8181 in their
browsers is when they are on
https://shibtest.dataverse.org:8181/shib.xhtml when creating or
converting an account after having authenticated with their institution.
Behind the scenes, other "plumbing" under "/Shibboleth.sso" is available
for metadata exchange and determining users' affiliation. A todo has
been added to make this port configurable from 8181.

Note that a servlet filter has been added because the same origin policy
was preventing resources from being read, which resulted in an ugly page
with missing images.

Finally, more information has been added when testing with TestShib to
not worry about the lack of the "mail" attribute.