Revert "Shib: use port 8181 for Shib page and plumbing #2117 "
This reverts commit 5c710d14c96a20eb027a3948657ef6534c0ae807.

This solution forces people installing Dataverse to run mod_shib on a
non-standard port (8181 rather than 443). At the very least this should
be configurable.

In addition it introduces a servlet filter (believed to be secure) to
get around the error "Cross-Origin Request Blocked: The Same Origin
Policy disallows reading the remote resource". This servlet filter
affects even installations that don't install Shibboleth and may
introduce a slight performance penalty.

Finally, we have reason to believe this solution will not work in
Harvard's production environment which includes multiple web servers
behind a load balancer ("content switch"). We haven't figured out a
configuration in staging that we believe will work.

All this is to say that this solution is non-standard and untested so it
doesn't belong code that we soon hope to tag with "4.0", which is why we
are reverting it.