Add simple tests for DataFileServiceBean
This adds test expecting that `null` DataFiles, DataFiles without content type
and DataFiles with a bogus content type (`"foo/bar"`) are *not* classed as files
of types that Dataverse handles in specific ways. Also, for these files the tests
expect that no thumbnail can be generated.

In my PR #2422 I changed the behaviour of `DataFileServiceBean.getFileClass` to
return `null` when the input is `null`. This is still discussed, so I don't test for
this behaviour. The same goes for `DataFileServiceBean.ingestableAsTabular`; this
method is not tested at all at the moment.

To make the tests pass, I changed the `isSpssPorFile` and `isSpssSavFile` methods
to return `false` when the input is `null`. Hopefully this makes sense to you too :)

@scolapasta, @landreev and @michbarsinai, I'll happily discuss these suggestions.