Sword Auth: use Dataverse 4 permission model #1070 #2495 #784
We are switching away from the old "admin only" permission checks we
were forced to used in the DVN 3.x days when the permission system was
less flexible. Now the rule is that if the permission system doesn't
prevent you from performing an operation (create dataset, upload file)
SWORD won't prevent you either.

Two SWORD operations had to be rewritten to support Dataverse 4
permissions.

First, retrieval of the Service Document with
permissionService.getDataversesUserHasPermissionOn has been changed to
pay attention to a user's groups. It's very common for the
":authenticated-users" group to be given a role that should allow SWORD
operations, for example.

Second, the "list datasets in a dataverse" operation
(listCollectionContents) was rewritten to allow anyone with `AddDataset`
permission to be able to execute the command to they can see if a
dataverse has been published or not but the only datasets that are
returned are those for which the user can issue `UpdateDatasetCommand`.
In the root dataverse, for example, the user should only see their
datasets.