Persist immutable id (a number) rather than mutable one #3338
The persistentuserid field in authenticateduserlookup now contains an
immutable value (i.e. "1938468") rather than a mutable one such as a
username (i.e. "jane_doe"). This is much preferred because most systems
allow you to change usernames (from "jane_doe" to "jane_smith", for
example, sometimes with side effects) in the case of divorce, etc.

Until #1445 is fixed, we'll persist the mutable value ("jane_doe") in
the useridentifier field in the authenticateduser column because it's
what is (unfortunately) displayed in the UI. This value is also used for
the assigneeidentifier field in the roleassignment table (see #1151).