API consistency: OK, response only has "data" #1612
Whenever the native API returns an OK response, it returns a JSON object
with the key "data" and lots of detail within. "data" is the only key
returned so I'm moving "message" to be within the "data" object to be
consistent with the rest of the native API.

I'm also removing the `okResponseGsonObject` method from the
AbstractApiBean for consistency. We don't want it to be cluttered with
Gson, Jackson or other methods.