remove experimental show_thumbnails boolean from Search API #3559
The show_thumbnails boolean in the Search API was used before we made
the decision to add an API endpoint to the Datasets API that returns an
InputStream. It was a way to enable automated API testing of the dataset
thumbnail. Now, we test InputStream instead as a well as a new "admin"
endpoint that provides us a look into the state of a dataset thumbnail.