add initial Swift support #3747
This branch was created based on on 4.5-export-harvest-swift which
recently had code from ferrys:4.5-export-harvest-swift-update merged
into it. This new branch was not a straight merge, however, because
there was some cruft in it that we don't want in the git history. The
latest from develop was merged into this new
3747-swift-and-compute-button branch and a `git reset` was performed to
make a single commit with all of the changes (and only the changes) that
we think we want. This single commit is being made by @pdurbin but thank
you to @ferrys @anuj-rt and @landreev who did all the actual work!