Changed dialog/documentation around alphabetical character rule #3150
Alphabetical is a superset of LowerCase & UpperCase, and if both are provided the code will count both as a rule in terms of the number of rules that need to be completed. The documentation now discourages admins from setting both, and the UI reflects that both are in place as we do not want to hide it.