Dataverse is too fat to run in 1 GB of memory #4040
Back out of the "kick the tires" story of using OpenShift Online instead
of Vagrant because Dataverse won't fit in the free tier which limits
your application to 1 GB of total memory. PostgreSQL and Solr each seem
to run fine in 256 MB of memory but Dataverse/Glassfish can't run in
512MB of memory. That said, logic has been added to the installer to
check if we're running in Docker and how much memory we have. We don't
change the Glassfish heap size based on this value, however, because the
war files fails to deploy.