revert eea4655 and tweak FileDownloadHelper to fix IP Groups #4268
eea4655 commented out the "Add permissions gained from groups" code so
we are re-enabling it. FileDownloadHelper was making an assumption that
a Guest user can never download a file. If you are using IP Groups, it
is a desired behavior for a Guest user to be able to download a file if
the user is part of an IP Group.