Attempt to clean http request via wrapper. Didn't really work $4204
I tried to wipe out all the parameters via HttpServletRequestWrapper, which is the documented way to alter query parameters. The problem is that just wiping out the query parameters throws errors with primefaces and possibly other places too. I even tried to preserve the parameters primefaces seemed to want to hold onto, but what those were wasn't obvious and I stopped walking down that rabbit hole.

The code as is does stop the logs from filling but if a user enters a malformed url they are just given a blank page.