QDR-901 Maintain the query part of passive login redirects
Before this fix, a passive login would not keep the query portion of a
redirect URL (required for dataverse search and privateurls, don't have
a specific drupal example). Adding this helped expose another flaw where
doing the passive check on any URL that had a path portion (e.g.
/discover/pilots) only set the _check_passive cookie for that subpath,
allowing the check to repeat when navigating to the home page (for
example) and, now with the query portion retained, causing an infinite
loop if the check fails (cookie was set on path/discover (for example),
check occurs and fails, ispassive then triggers on the error url (of the
form /?now..., sets the cookie to that URL, and then, for all subsequent
calls, it requests a redirect to that same location).

The two changes to add the query portion and set the path to / allow
supporting passive login to urls with queries (without an infinite loop)
and fix the pre-existing issue that going to a page at a subpath would
result in additional unnecessary passive checks for / or other paths as
they were accessed.