QDR-901 - handle URLs with ;
; is not an allowed char in cookie values,
but we see it in urls on beta - where a ;jsessionid=... is being
inserted between the path and query. encoding and decoding going in/out
of the cookie handles it. (before this fix, our parsing assumed the
cookie value stopped at the end of the path (seeing the ; in front of
jsessionid), giving the same symptom as not sending the query part at
all...)