saving from all pages works #4343
I originally tried to take the approach of saving the context multiple times and passing the result around, but could not wrap my head around how to do that correctly with all of our pages. Instead I took the approach of only saving the context once by having a toggle for the save context in the commands.