switch time-based queries from 12 months to 1, no running total #4527
These time-based queries now simpy return a count, a total of the number
of datasets or downloads or whatever. You specify the month in "YYYY-MM"
format.

There are FIXMEs in the code to always return the current month.

There are also FIXMEs in the code about preventing SQL injection
attacks.