Prov: Fix save context case #4343
The logic in the previous commit was incorrect. Turns out the bug was due to dataFile attributes only being saved if an upload action was taking place on editDatafilePage. Now the prov code only saves the datafile if it isn't being saved by upload.

Also, the provenance freeform saving was updated so that it wasn't triggering saving in unneeded cases