The "contents" API now requires a superuser API token? #4675
It's unclear what changed but for the purposes of getting the
API tests in SwordIT passing, we are switching from a contributor
API token to a superuser API token when calling the "contents" API.
A TODO has been added to investigate this futher.

Clarifying comments and FIXMEs have been added as well. The contents API
is known to be non-performant (#2122).