remove doiSeparator key after merge with DOI for files code
Both Handles and DOIs require '/' as the authority/identifier separator.
Since any separator / sub-identifier separator is now just part of the
shoulder, there is no need for a configurable value for that. It is also
problematic to make the authority/identifier separator configurable at
this point since there are many places in the code where '/' is
hardcoded (reasonable given that Handle and DOI both require it), so
supporting a new Id type that doesn't use this character requires more
work than creating a new serviceBean and setting the key.

Nominally, all identifiers will have '/' as the separator in the
database and the separator column in the dataset table (and now for
datafiles too) could be removed, but that would complicate id matching
relative to current code which just concatenates
<authority><separator><identifier> and would add work to restore it if a
different identifier with new separator is ever created. So - leaving
this, and code to get/set it as is...