Separating the choice of identifier generation style for datasets and files
Given the code for generating datafile ids, it looks like the new
ability to support a shoulder should only apply to datasets as a start.

This commit splits the existing IdGenerationStyle key into two one for
datasets and one for datafiles. Datafiles will continue to only support
random or sequential identifiers (with an additional choice as to
whether to those are independent or dependent (appended to the dataset
identifier) while datasets can also use the shoulder with random or
sequential options.

It seems like allowing files to use a shoulder when they are independent
would also make sense, but I don't think using that with dependent would
make as much sense. If that's a reasonable choice, we might want to
rethink having the dependent/independent choice as a separate key
(datafiles would ultimately have 6 choices - random or sequential with
dependent/independent or independent with shoulder and random or
sequential).