shoulder support for datafiles and fix for uniqueid checks with shoulder
These updates fix an issue with checking the db for entries that match
the new one so that tests include the shoulder.

They also use a shoulder, if defined, for datafile, only when the ids
for datafiles are being generated as 'INDEPENDENT'. When DEPENDENT, the
shoulder is already used once in the part of the id coming from the
dataset, so it is not used again.