avoid second exception
when testing for #3328, the call to
idServiceBean.getProviderInformation() in the catch clause causes an
javax.ejb.EJBException due to a null pointer exception that appears to
be caused by an earlier javax.ejb.TransactionRolledbackLocalException
when the 400 response is received from datacite.

One way to avoid this is to retrieve the provider information prior to
any exceptions. I've verified that this results in the
dataset.publish.error being shown.