don't run if dataset locked for pidRegister
have seen this. Nominally this should not run until the publication
process is done and pidRegister lock removed, so perhaps there's a race
condition. At a minimum, we don't want to run the archiving if
pidRegister is not done, so we can fail. If it happens often, we can fix
or wait, etc.