4990 working script first pass
Script now fufills the need of creating an ec2 instance and scping in a file that has the branch name to be spun up later by ansible.

This could probably be used as is, but I want to clean it up and create a tear-down script or think of some other ways we will ensure that a billion instances haven't been created.