Add direct dependency to Jackson JSON-B and JSON-P in pom.xml.
Needed because the transitive dependency from AWS has been used before.
Thus this relates to #5274. As best practise, code should not rely on
those deps but add a direct dependency.

For the sake of working on #4260 (Java EE 8) at some point in the future, please
remember to refactor the code using Jackson (and Gson) and remove it
in favor of Java EE 8 native JSON-B and JSON-P support.