deduplicate permissions
Original code could have duplicates from RoleAssignments for different
roles containing the required perm and/or from different definition
points in the parent hierarchy. Rather than keep duplicates (and then
have to cache the expensive step of finding the user or group) now using
a Set of ids and then creating a list with unique entries.