Don't rely on window.opener
Looks like the terms of access dialog manages to open the ATI window
without setting window.opener, so we need to check for the null. This
means that the 'return to dataset' button goes back to the dataset and
not the file when clicking view on the file page and going through the
terms dialog.