#5674 Changed GET metadata format to one used for add/replace
This involved fixing the implementation on the FileMetadata GSON deserializer. It is not 100% done as some params are not marked for serialization but mostly there.
This also includes a unit test and an IT test around the functionality.
I also pulled in the additions needed to act on the DirectoryLabel