dcm - use new option name - assume store with id 's3'
Note: DCM doesn't use any of the other S3 options in creating the Amazon
 client, so it is somewhat hardcoded that was as well. A rewrite might
 identify a specific s3 store to use, build the amazon client from it's
settings and still use the general dcm-s3-bucket-name entry. If DCM
should be supported across multiple stores, a
dataverse.files.<id>.dcm-s3-bucket-name option could be created so that
everything is ties to a specific store.