only set the create time if createIdentifier succeeded #5093
The fix required removing the throwing of RuntimeException in
DataCiteRESTfullClient. This is considered a system exception and rolls
back the entire "create dataset" command and transaction.

This commit also remove the temporary fix I had put in to allow datasets
to be created when DataCite is down: 9cbd5d3

Basically, it's a better fix. We need the create time to be null if
DataCite is down so that we can query for unreserved DOIs and then
reserve them.