Update codebook2-0.xsl
Replacing a hardcoded value with an XPath acquired value. I have made a presumption that the distributor name is the appropriate element to present here, since it is paired with the DOI value. This is my first Dataverse pull request, so let me know if I am missing a step or whatever.