change to min-part-size
it's the max that a single part direct upload can be (as coded), and the
minimum part size for multipart uploads.
These could be separate settings, but it's not clear that that's useful
to support. E.g. for AWS S3, the min-part-size is 5MB but single uploads
could go to 5 GB so that is the max single upload.