Provide @DataSourceDefinition for app-level DB config. #6819
This commit is a first test to apply database configuration and
credentials to the application server without using a config
on domain level but on application level. It is not very configurable yet.

The JNDI names have been changed to be conform with Java EE 7.
See https://github.com/javaee-samples/javaee7-samples/tree/master/jpa/datasourcedefinition-applicationxml-pu
and others for working examples. (Staying with the old name was not successfull.)

We had to use the "global" scope, so the persistent EJB timers are successfully created
and stored in the database. Using the "app" scope crashes the deployment.