bump version of postgres (within 9.6), clarify docs #7241
We got a question about all the pgAdmin stuff. What it's for, what's
going on. It wasn't clear that we're simply using this tool to check
if the "trust" change worked (which simplifies the installation later).
Perhaps in the future we could switch to a test with `psql` but an
advantage of mentioning pgAdmin early is to let the new developer know
that this tool exists.