remove binding and process=@this
Our goal is to include file-download-popup-fragment.xhtml onto another
location, the Preview tab on the file page, while still displaying
it as a popup on that page.

In the Preview tab the input fields weren't showing up and we discovered
that commenting out "binding" made them appear. Commenting out "binding"
resulted in validation not working. Additionally removing process=@this
along with much more refactoring, got validation working. Bean validation
was added as well.

The hope is that with this commit in place, we can switch back to trying
to get the extra "include" (mentioned above) working.

Some things to note about this commit:

- Custom UIInput fields have been removed in favor of standard
Bean Validation.
- The GuestbookResponse field was removed from FileDownloadHelper because
it was confusing to sometimes set it and other times pass it in as argument
to various methods. Now we always pass it in.
- Various unused code was deleted.
- process=@this wasn't removed from the "package" code and it's unknown
if that code is working.