fix for edu.harvard.iq.dataverse.api.AdminIT test fail in Java 11
The DV code tested in testLoadMetadataBlock_ErrorHandling assumed it
could parse the message of an ArrayOutOfBounds exception as an it to
determine the column that fails. This message is now a String. Rather
than parse it (and fail if it changes), I modified the code so that the
length of the values array is visible in the catch and can be sent
directly (the first out of bounds index is if/when the index is
values.length).