user merge and delete cleanup, get traces #4475 #7575
Primarily this has been an investigation into existing code for
deleting and merging users. Additionally, a "get user traces"
command was added to get a sense of why a user can't be deleted
or what would be merged. The top of DeleteUsersIT has a lengthy
comment about which database tables are in play and a variety
of scenarios involving users to be deleted, merged or (in the
future) disabled (#2419).

The following bugs where fixed:

- Unable to merge (or delete) a user when they have initiated
a password reset. #7575
- Delete OAuth2 tokens on delete.