restricted summary stats #7619
- At the dataset level, DDI exports no longer show "dataDscr" information
for restricted files. There is only one version of this export and it is
the version that's suitable for public consumption with the "dataDscr"
information hidden for restricted files.
- Similarly, at the dataset level, the DDI HTML Codebook no longer shows
"dataDscr" information for restricted files.
- At the file level, "dataDscr" information is no longer publicly
available for restricted files. In practice, it was only possible to
get this publicly via API (the download/access button was hidden).
- At the file level, "dataDscr" (variable metadata) information can still
be downloaded for restricted files if you have access to download the file.