more efficient check for disabled or deleted users #2419 #4475
Previously, in 3e59e70 we added logic to check if the user in the
session has been deleted or deactivated. The problem is that this code
is called often so now we only do this check if the user tries to go to
the account page or tries to execute a command.