revert(config): make web.xml variable substitution compatible with Payara 5.2020.6 #6953
Reverts commit ca354119

Due to the nature of the application server, when the web app context is
initialised, it cannot yet access the deployments default MPCONFIG
source within META-INF/microprofile-config.properties.

As a consequence, as long as we do not update to Payara 5.2021.1, we
cannot provide sane defaults within the variable. That means that every
installation would need to set the variables in MPCONFIG sources
available at startup time and before the deployment finished: any
are possible that are baked into Payara.

As this doesn't seem feasible, this PR will be blocked until #7700 is
resolved.