Don't reuse HttpClientContext
There are nonthreadsafe elements in it. Since we don't use cookies
currently in the Dataverse API, I'm removing it. For the general case,
e.g. if you can login via API and use a cookie to maintain an api
session, we'll need to add a per-thread context back (sharing the cookie
store)