Don't call archiver directly
The direct call here was originally from when the Curate command was
used (in which workflows don't run but any current archival copy needs
to be updated). Now that FinalizePub is used instead, it will run any
post-publication workflow (e.g. for archiving) and/or the admin can run
archiving separately.