feature(mpconfig): introduce the concept of scopes and make the `SettingsServiceBean.Key` enum use of it. #7639
By adding these scopes as a data structure, we enable a documented approach to define
key names inside of MicroProfile Config Sources. Instead of a collection of Strings,
this is much more explicit. Due to its iterable nature, it may be used to generate documentation
down the road. Changing names gets more trackable.