fix(jpa): introduce real defaults for DataSourceDefinition via MPCONFIG. #7980
Previously, with Dataverse software 5.3, the option to configure
the database connection has been moved into the codebase. Admins
can set details via MicroProfile Config.

With updating to Payara 5.2021.4, we can provide default values
for the connection details. Before, this had been tried with adding
them to META-INF/microprofile-config.properties. However, this is
not possible due to the timing of resource creation in the application
server vs. reading the properties file.

IQSS/dataverse#7980