build(deps): make package nu.xom.xom a managed dependency IQSS/dataverse-security#38
The XOM library had been changed to com.iom7.xom before, trying to
achieve updated dependencies. It had been declared as a direct
dependency, which it is not - the library is nowhere used in the
Dataverse codebase. Instead, it is used by sword2-server and others.

As the original authors of XOM added never versions, so this commit
removes the direct dependency and adds a newer version to
<dependencyManagement>.

This has a sideeffect: now the xercesImpl library version is defined by
Apache Tika (depending on it) instead of Xom (which is a sub-dependency,
making the relying version there loose against Tika).

This update of xercesImpl is necessary as the old version (the code used
<=2.8.0) contains security issues fixed in 2.12.1, which Tika depends
on.

It does not seem necessary to add xercesImpl to <dependencyManagement>
yet. That might change in the future to avoid drifting versions...

See IQSS/dataverse-security#38