refactor(branding): make BrandingUtilTest reusable and JUnit5-ish
Some other tests might rely on the BrandingUtil to actually do things.
This refactoring ensures proper mocking and cleaning up after the tests
ran, trying to ensure better atomicity.

It enables easy setup and teardown of mocks to be used in other tests.
It might be necessary to expose the underlying mocked objects for deeper
testing. Left for later exercise if ever necessary.

This refactoring also
- makes a few tests parameterized to be easier to read,
- enables much easier identification of failing test examples and
- removes static ordering now unnecessary.