exclude embargoed files when autoselecting a thumb.
Nominally one could also check whether emargoed files are still
embargoed with some additional query logic. However, to avoid the
expense, and, since this method just goes ahead and assigns the file it
finds as the dataset's thumbnailfile, the autodetect should only run
once, soon after publication, when, presumably, all the embargoes were
in force. After that, the only way to change the thumb is to via the
edit/thumbs and widgets dialog and there, we have the full logic to list
files who's embargoes are over. So - while this could theoretically
result in a dataset not using a valid file for its thumb, in practice it
shouldn't happen afaik.