Update harvestserver.rst
Tried to clarify that the path should include all parent collections up to but excluding the Root collection.

Also added @pdurbin's suggestion of using an API endpoint to get the Dataverse collection's database ID. It might be helpful to link to the section of the API Guide that documents that endpoint, which I think is https://guides.dataverse.org/en/5.7/api/native-api.html?highlight=unlock#view-a-dataverse-collection, but I don't know the best way to do this in this rst file. Tried looking through this rst file to for an example but I'm still not sure. @pdurbin could you add a link behind ""View a Dataverse Collection" API endpoint"?