remove role="menu"
per https://web.archive.org/web/20190116170516/https://www.marcozehe.de/2018/09/22/wai-aria-menus-and-why-you-should-generally-avoid-using-them/
- it doesn't seem like we're using this role as intended in most cases,
so I'm doing a blanket removal to see if there are adverse UI impacts or
new accessibility issues being reported. With this role, menus like the
Edit menu in the files table report an issue (from AXE) with the menu
not containing child elements with the right role (i.e. menuitem), and a
seemingly spurious 'li elements must be contained in a ul or ol' error.