Update harvestserver.rst
From our edits to the section about the subtreePaths query, I copied the extra tip about using an API endpoint to get the database ID to the parentId query. That's the only other place, at least on this page, where the user needs to know what the database ID is and is given tips for how to get it.