dataset files table - unique titles for preview links
Links with the same name should have similar purpose

This change uses the i18n names of individual previewers for the preview
button, which reduces the ambiguity, but doesn't give the preview in
each row a unique title. (I.e. two images still have 'View Image' as the
title, but 'Read Text' for a text preview would now be different.)

We could add the filename to the title to make this fully unique, either
with/without using the previewer names.

Note: Deque says this is an issue unless the links are 'intentionally
ambiguous' - not sure if that includes assuming the row context is
enough.