get license selections from service, add caching, clear in terms
Together these changes remove the last parts of dataset-license that
relied on the DatasetPage for license selection which allows templates
to work correctly with multilicense. Changes here include

getting license selections from the LicenseServiceBean
removing code for caching license state and selections from DatasetPage
caching the list of licenses in SettingsWrapper
clearing the terms that aren't allowed with a non-null/non-custom
license in the TermsOfUseAndAccess class itself as part of setting the
license
minor optimization in the same class to not copy those terms if there is
a non-null license.