Fix error response/tests
Runtime errors returned from a bean trigger a transaction rollback
exception and either those exceptions have to be caught (i.e. as in
https://github.com/IQSS/dataverse/blob/e7be5947abf2fbaa9ceaac09b27a60d439647f24/src/main/java/edu/harvard/iq/dataverse/search/SearchServiceBean.java#L354-L357)
or some non-runtime exception has to be returned. I used the
WrappedResponse class as a way to simplify but reviewers may have some
other preference (e.g. just catch the transaction exceptions as in the
SearchServiceBean or use non-runtime exceptions directly in the
LicenseServiceBean class). Any of these should assure a non-null
response is returned which in turn should allow the AdminIt tests to
pass.