fix replace file failure
There is weirdness here. In one use, the fileMetadata.label isn't added
to the id (the id is created before the fileMetadata param is set) while
in another (i.e.g in the file/replace use case) it is set adn causes a
failure because jsf doesn't allow a '.' char in the id (which is part of
many filename.ext labels). Since this isn't needed for accessibility (as
the id is unique due to primefaces prefixing it), I can just revert to
not having it in the id. (I originally added it since we do need it in
editFilesFragment to assure unique ids in the file table (where a div is
used instead of p:outputPanel) and I was trying to have analogous names
here)