refactor(zipdownloader): replace Maven Assembly Plugin with Spring Boot #8394
The Spring Boot method to create an executable fat/uber JAR is offering a much better
experience than what Maven Assembly does. It provides a custom bootloader to load
JARs from the JAR, which means less potential for conflicts etc.

It also offers options to be integrated in other workflows like container images etc.
The package is now about 8.2 MB in size, which is negligible compared to 8.0 MB via
the Assembly plugin.