refactor(deps): move SolrJ dependency to parent POM #7662
As the SolrJ dependency is going to be used within more than the WAR module,
it is getting moved to the parent POM.

In the same go, the dependency is versioned with a property "solr.version",
that makes it easy to switch to a different Solr version.

The property has been present before, but now is updated to 8.11.1 to
reflect the latest upstream changes.

The Sphinx guide conf.py is also updated to retrieve the property from
the parent POM instead of the WAR POM.