feat!(solr): create Maven module solr-configset #7662
With this commit, a new Maven submodule is introduced: creating a
distributable configset to create Solr cores from.

The configset consists of a customized core configuration, a schema and
more to create indices from incoming Solr docs.

As the solrconfig.xml file depends on the Solr/Lucene version in use
under the hood and requires careful tweaking for usage within a
Dataverse installation, it gets created by a compilation process
involving version controlled XSLT files.

The schema.xml file is version controlled and depends on the application code.

A Solr release version according to the parent POM defined Solr version
is downloaded and cached with the local Maven repository. The template
for out Dataverse flavored configset is extracted from it.

With the resulting configset an embedded Solr instance is started and a
core gets created with it to ensure basic functionality of a compile
Solr configset and schema.xml.

BREAKING CHANGE: :warning: changes Solr configuration method from
replacing single files to creation from a template