feat(solrteur): introduce state machine for TSV file layout #7662
Our custom metadata block TSV files follow a certain order of things.
We also do not allow for repetitions or similar. All of this can be
most easily be depicted with a state maschine, so we know where to send
a line to for parsing.

This commit also adds the very basic (empty) POJOs to store the block,
fields and vocabularies in to enable testing the state transition.

It also adds constants we rely on, like what's the trigger char, the
comment intro and the field delimiter