feat(solrteur): introduce a validator helper class #7662
The TSV parser needs to verify if a certain line is a header line
and matching the spec. To avoid duplicated validation code,
this validator can be used with an arbitrary list of strings
(so it can be reused for blocks, fields and vocabularies).

As we will need to validate URLs in certain fields, this validator
also offers a helper function to create predicates checking for valid URLs.