feat(solrteur): extend the metadata block POJO #7662
The Block POJO now contains the header specification (uses the Validator class
to perform the validation) and allows to parse a line into a List.
A later relaxation of the spec allowing for reordering of fields, etc is possible,
while the calling code of the parser can reuse the found header definition.

A builder pattern is used to parse and validate the actual definition.
As the block may only be used once the definition, all fields and vocabularies
have been parsed (if the is an error within the TSV the parsing has to fail!),
the builder pattern is a natural match to that.